backend/requirements.txt
# Prima Scholar Backend Dependencies

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.21
cryptography==41.0.4

# AI and ML
openai==1.3.5
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2

# Document Processing
PyMuPDF==1.23.5
python-docx==0.8.11
pypdf==3.16.2
textract==1.6.5

# Caching and Performance
redis==5.0.0
python-memcached==1.59

# API and Serialization
marshmallow==3.20.1
requests==2.31.0
python-dotenv==1.0.0

# External Tools Integration
sendgrid==6.10.0
google-api-python-client==2.100.0
slack-sdk==3.22.0
python-linkedin-v2==2.0.1a0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
Werkzeug==2.3.7

# Testing (Development)
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Security
bcrypt==4.0.1
PyJWT==2.8.0
itsdangerous==2.1.2

# Monitoring and Logging
structlog==23.1.0
colorlog==6.7.0

# File handling
python-magic==0.4.27
Pillow==10.0.1

# Math and Statistics
pandas==2.0.3
scipy==1.11.2

# Development Tools
black==23.7.0
flake8==6.1.0
isort==5.12.0